+++ ios/RCTDBModule.h
@@ -1,0 +1,16 @@
+//
+//  RCTDBModule.h
+//  ScrabbleHelper
+//
+//  Created by Adam Kruczek on 03/04/2023.
+//
+
+#import <React/RCTBridgeModule.h>
+
+#ifndef RCTDBModule_h
+#define RCTDBModule_h
+
+@interface RCTCalendarModule : NSObject <RCTBridgeModule>
+@end
+
+#endif /* RCTDBModule_h */