+++ ios/RCTDBModule.m
@@ -1,0 +1,15 @@
+//
+//  RCTDBModule.m
+//  ScrabbleHelper
+//
+//  Created by Adam Kruczek on 03/04/2023.
+//
+
+#import <Foundation/Foundation.h>
+#import "RCTDBModule.h"
+
+@implementation RCTDBModule : NSObject
+
+RCT_EXPORT_MODULE(DBModule);
+
+@end