+++ ios/RCTDBModule.swift
@@ -1,0 +1,8 @@
+//
+//  RCTDBModule.swift
+//  ScrabbleHelper
+//
+//  Created by Adam Kruczek on 03/04/2023.
+//
+
+import Foundation