+++ ios/RCTEventEmitter.h
@@ -1,0 +1,18 @@
+//
+//  RCTEventEmitter.h
+//  ScrabbleHelper
+//
+//  Created by Adam Kruczek on 03/04/2023.
+//
+
+#ifndef RCTEventEmitter_h
+#define RCTEventEmitter_h
+
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
+
+@interface EventEmitter: RCTEventEmitter <RCTBridgeModule>
+
+@end
+
+#endif /* RCTEventEmitter_h */