+++ ios/RCTEventEmmiter.m
@@ -1,0 +1,15 @@
+//
+//  RCTEventEmmiter.m
+//  ScrabbleHelper
+//
+//  Created by Adam Kruczek on 03/04/2023.
+//
+
+#import <Foundation/Foundation.h>
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
+
+@interface EventEmmitter : RCTEventEmitter <RCTBridgeModule>
+
+@end
+