+++ src/native-db/native-db.models.ts
@@ -1,0 +1,6 @@
+import { NativeModulesStatic } from 'react-native'
+
+export interface NativeDB {
+  findPossibleWords: (allWords: string[]) => string[];
+  _nativeModule: NativeModulesStatic;
+}